--- .github/dependabot.yml
@@ -1,20 +1,0 @@
-version: 2
-updates:
-- package-ecosystem: pip
-  directory: /
-  schedule:
-    interval: daily
-  labels:
-  - enhancement
-- package-ecosystem: github-actions
-  directory: /
-  schedule:
-    interval: daily
-  labels:
-  - enhancement
-- package-ecosystem: docker
-  directory: /
-  schedule:
-    interval: daily
-  labels:
-  - enhancement
