--- .github/workflows/quality.yml
@@ -1,30 +1,0 @@
-name: Quality
-
-on:
-  push:
-
-jobs:
-  pre-commit:
-    runs-on: ubuntu-latest
-    steps:
-    - name: Checkout
-      uses: actions/checkout@v3
-
-    - name: Set up Python
-      id: setup-python
-      uses: actions/setup-python@v4.5.0
-      with:
-        python-version: '3.11'
-
-    - name: Run pre-commit
-      uses: pre-commit/action@v3.0.0
-
-  summary:
-    runs-on: ubuntu-latest
-    needs:
-    - pre-commit
-
-    steps:
-    - name: Generate Summary
-      run: |-
-        echo '### Hello world! :rocket:' >> $GITHUB_STEP_SUMMARY
